Rewrote events
Now starting the rewrite of the event system for the implementation of
the rune interpreter.
- Split EventListener into InventoryListener and DropListener
